Separated test for custom issue
