+++ tests/test_run.py
@@ -11,13 +11,17 @@ from cleanvision.issue_managers.image_property import BrightnessProperty
 from cleanvision.issue_managers.image_property_issue_manager import (
     compute_scores_wrapper,
 )
-from docs.source.tutorials.custom_issue_manager import CustomIssueManager
+
+
+@pytest.fixture()
+def imagelab():
+    data_path = os.path.join(os.path.dirname(os.path.abspath(__file__)), "./data")
+    imagelab = Imagelab(data_path=data_path)
+    return imagelab
 
 
 @pytest.mark.usefixtures("set_plt_show")
-def test_imagelab_run(capsys):
-    data_path = os.path.join(os.path.dirname(os.path.abspath(__file__)), "./data")
-    imagelab = Imagelab(data_path=data_path)
+def test_imagelab_run(capsys, imagelab):
     imagelab.find_issues()
     imagelab.report()
     captured = capsys.readouterr()
@@ -49,6 +53,11 @@ def test_imagelab_run(capsys):
     assert "Issues found in images in order of severity in the dataset" in captured.out
     assert "Number of examples with this issue:" in captured.out
     assert "Examples representing most severe instances of this issue:" in captured.out
+
+
+def test_get_stats(imagelab):
+    stats = imagelab.get_stats()
+    assert stats == imagelab.info["statistics"]
 
 
 def test_incremental_issue_finding(generate_local_dataset, len_dataset):
@@ -105,20 +114,6 @@ def test_incremental_issue_finding(generate_local_dataset, len_dataset):
             "brightness_perc_1",
         ]
     ) == set(imagelab.info["light"].keys())
-
-
-@pytest.mark.usefixtures("set_plt_show")
-def test_custom_issue_manager(generate_local_dataset, len_dataset):
-    imagelab = Imagelab(data_path=generate_local_dataset)
-    issue_name = CustomIssueManager.issue_name
-
-    issue_types = {issue_name: {}}
-    imagelab.find_issues(issue_types)  # check for custom issue type
-
-    assert set(imagelab.issue_summary["issue_type"].values) == set([issue_name])
-    assert len(imagelab.issues) == len_dataset
-    assert set(["is_custom_issue", "custom_score"]) == set(imagelab.issues.columns)
-    assert set(["statistics", "custom"]) == set(imagelab.info.keys())
 
 
 @pytest.mark.parametrize(
